implement getAll functionality for reviews and tours using factory methods
