+++ controller/handlerFactory.js
@@ -1,5 +1,6 @@
 const catchAsync = require('./../utils/catchAsync');
 const AppError = require('./../utils/appError');
+const APIFeatures = require('./../utils/apiFeatures');
 
 exports.deleteOne = (Model) =>
   catchAsync(async (req, res, next) => {
@@ -60,6 +61,33 @@ exports.getOne = (Model, popOptions) =>
       status: 'success',
       data: {
         data: doc,
+      },
+    });
+  });
+
+exports.getAll = (Model) =>
+  catchAsync(async (req, res, next) => {
+    //! HACK - to also work with reviews
+    let filter = {};
+    if (req.params.tourId) filter = { tour: req.params.tourId };
+
+    // console.log(req.query);
+    //? BUILD QUERY OBJECT
+    const features = new APIFeatures(Model.find(filter), req.query)
+      .filter()
+      .sort()
+      .limitsFields()
+      .paginate();
+
+    //? EXECUTE QUERY
+    const docs = await features.query;
+
+    //? SEND RESPONSE
+    res.status(200).json({
+      status: 'success',
+      results: docs.length,
+      data: {
+        data: docs,
       },
     });
   });
