+++ controller/reviewController.js
@@ -2,27 +2,14 @@ const catchAsync = require('./../utils/catchAsync');
 const Review = require('./../models/reviewModel');
 const factory = require('./handlerFactory');
 
-exports.getAllReviews = catchAsync(async (req, res, next) => {
-  let filter = {};
-  if (req.params.tourId) filter = { tour: req.params.tourId };
-  const reviews = await Review.find(filter).select('-__v');
-
-  res.status(200).json({
-    status: 'success',
-    results: reviews.length,
-    data: {
-      reviews,
-    },
-  });
-});
-
 exports.setTourUserIds = (req, res, next) => {
   if (!req.body.tour) req.body.tour = req.params.tourId;
   if (!req.body.author) req.body.author = req.user._id;
   next();
 };
 
-exports.getReview    = factory.getOne(Review);
+exports.getAllReviews = factory.getAll(Review);
+exports.getReview = factory.getOne(Review);
 exports.createReview = factory.CreateOne(Review);
 exports.deleteReview = factory.deleteOne(Review);
 exports.updateReview = factory.UpdateOne(Review);
